backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "openai>=1.50.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
    "anthropic>=0.64.0",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "langgraph>=0.6.6",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "fastembed>=0.7.3",
]
